Cast ADC_VREF to double (be specific)
- http://www.avrfreaks.net/index.php?name=PNphpBB2&file=viewtopic&p=998242#998242

+++ board.h
@@ -27,7 +27,7 @@ void init_board(void);
 
 static inline double cnv2volt(uint32_t cnv)
 {
-	return cnv * (ADC_VREF / (1UL << ADC_BITS));
+	return cnv * ((double) ADC_VREF / (1UL << ADC_BITS));
 }
 
 #endif
